web/requirements.txt
# ECG Digitizer Server Dependencies
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
scipy>=1.9.0
scikit-image>=0.19.0
opencv-python-headless>=4.5.0
PyYAML>=6.0
gunicorn>=20.1.0
matplotlib>=3.5.0
